@@ -2,10 +2,9 @@ package com.example.wy521angel.slidetest;
 
 import android.app.ListActivity;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListAdapter;
 import android.widget.ListView;
@@ -13,7 +12,7 @@ import android.widget.ListView;
 public class MainActivity extends ListActivity {
 
     private String[] labels = new String[]{"layout方法，视图坐标方式", "layout方法，绝对坐标方式", "offset方法",
-            "LayoutParams方法", "scrollBy方法", "scroller方法","TwoViewPager测试"};
+            "LayoutParams方法", "scrollBy方法", "scroller方法","TwoViewPager测试","两个ScrollView嵌套的滑动冲突"};
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -22,6 +21,7 @@ public class MainActivity extends ListActivity {
                 labels);
         getListView().setAdapter(adapter);
     }
+
 
     @Override
     protected void onListItemClick(ListView l, View v, int position, long id) {
@@ -47,6 +47,9 @@ public class MainActivity extends ListActivity {
                 break;
             case 6:
                 startActivity(new Intent(this, Activity0.class).putExtra("label", 6));
+                break;
+            case 7:
+                startActivity(new Intent(this, Activity0.class).putExtra("label", 7));
                 break;
         }
     }
